Refactor TensorFlow.js tensor operations and improve type safety
- Updated tensor manipulation methods to use `tf.reshape` and `tf.cast` for better clarity and type safety.
- Replaced deprecated methods like `as3D`, `as4D`, and `toFloat` with their respective alternatives.
- Enhanced the handling of tensor shapes in various neural network classes, ensuring consistent tensor dimensions.
- Improved error handling and type definitions in test files for better maintainability.
- Added support for WebGL and CPU backends in test configurations.
- Updated TypeScript configuration for better compatibility with Node.js and modern JavaScript features.
- Introduced a new test runner configuration for streamlined testing with Playwright.

+++ src/mtcnn/Mtcnn.ts
@@ -76,7 +76,9 @@ export class Mtcnn extends NeuralNetwork<NetParams> {
     const tsTotal = Date.now();
 
     const imgTensor = tf.tidy(() =>
-      bgrToRgbTensor(tf.expandDims(tf.browser.fromPixels(inputCanvas)).toFloat() as tf.Tensor4D)
+      bgrToRgbTensor(
+        tf.cast(tf.expandDims(tf.browser.fromPixels(inputCanvas)), 'float32') as tf.Tensor4D
+      )
     );
 
     const onReturn = (results: MtcnnResult[]): { results: MtcnnResult[]; stats: MtcnnStats } => {
